Changed int to smallint in CreateDB, and finished??? Queries.sql

+++ CreateDB.sql
@@ -7,7 +7,7 @@ USE Streamflix;
 CREATE TABLE titles (
 	id INT NOT NULL AUTO_INCREMENT,
     title VARCHAR (255), 
-    year INT NOT NULL, 
+    year SMALLINT NOT NULL, 
     PRIMARY KEY (id)
     );
